// github/types/paths/repos/{owner}/{repo}/issues/{issue_number}/labels.types.ts
export type issuesListLabelsOnIssue = (
  $: OmitValueWhenNever<{
    query: issuesListLabelsOnIssue_Query;
    querystring: never;
    path: issuesListLabelsOnIssue_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: {
          Link: { schema: string };
        };
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: Array<label>;
          };
        };
        examples: {
          default: unknown;
        };
      };
      301: moved_permanently;
      404: not_found;
      410: gone;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Adds labels to an issue.
 */
export type issuesAddLabels = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: issuesAddLabels_Path;
    headers: never;
    cookie: never;
    body:
      | {
          /**
           * The names of the labels to add to the issue's existing labels. You can also pass an `array` of labels directly, but GitHub recommends passing an object with the `labels` key. To replace all of the labels for an issue, use "[Set labels for an issue](https://docs.github.com/rest/issues/labels#set-labels-for-an-issue)."
           */
          labels?: Array<string>;
        }
      | Array<string>
      | Array<{ name: string }>;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: Array<label>;
          };
        };
        examples: {
          default: unknown;
        };
      };
      301: moved_permanently;
      404: not_found;
      410: gone;
      422: validation_failed;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Removes any previous labels and sets the new labels for an issue.
 */
export type issuesSetLabels = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: issuesSetLabels_Path;
    headers: never;
    cookie: never;
    body:
      | {
          /**
           * The names of the labels to set for the issue. The labels you set replace any existing labels. You can pass an empty array to remove all labels. Alternatively, you can pass a single label as a `string` or an `array` of labels directly, but GitHub recommends passing an object with the `labels` key. You can also add labels to the existing labels for an issue. For more information, see "[Add labels to an issue](https://docs.github.com/rest/issues/labels#add-labels-to-an-issue)."
           */
          labels?: Array<string>;
        }
      | Array<string>
      | { labels?: Array<{ name: string }> }
      | Array<{ name: string }>
      | string;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: Array<label>;
          };
        };
        examples: {
          default: unknown;
        };
      };
      301: moved_permanently;
      404: not_found;
      410: gone;
      422: validation_failed;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Removes all labels from an issue.
 */
export type issuesRemoveAllLabels = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: issuesRemoveAllLabels_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      301: moved_permanently;
      404: not_found;
      410: gone;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type issuesListLabelsOnIssue_Query = {
  per_page?: number;
  page?: number;
};

export type issuesListLabelsOnIssue_Path = {
  owner: string;
  repo: string;
  issue_number: number;
};

export type issuesAddLabels_Path = {
  owner: string;
  repo: string;
  issue_number: number;
};

export type issuesSetLabels_Path = {
  owner: string;
  repo: string;
  issue_number: number;
};

export type issuesRemoveAllLabels_Path = {
  owner: string;
  repo: string;
  issue_number: number;
};